metric: add metric-v1 adapter as a separate version
Split the Metric DEX into two version adapters sharing one builder
(aerodrome-style shared file in the base folder):

- dexs/metric/utils.ts: getMetricAdapter(factory, chainConfig, events)
  builds the pool-discovery + volume pipeline once; event ABIs are
  injected per version.
- dexs/metric (legacy factory) keeps the default legacy event ABIs.
- dexs/metric-v1 (v1 factory) passes the v1 Swap/PoolCreated ABIs, which
  differ from legacy (int256 amounts, indexed tokens, poolAddress field).
